Refactor src/chart/utils
